fix(#297): 可空 dtype 的 code 欄不再讓解碼炸掉
decode_weight_keys 對 Int64/Float32 這類還帶著 pd.NA 的可空 dtype，直接 to_numpy(float64)
會 raise ValueError，而不是退回權重 1.0。weight 這條路徑全程是 graceful 的（欄位不存在、
表是空的都只記 log），不該是唯一會中斷訓練的那一環。

_weight_keys_and_codes 的 dropna=False 註解本來就寫著「Int64 + pd.NA 在這裡可達」，
所以這不是理論上的邊界。先 astype("float64") 再 to_numpy 就好。

一併把實際見過的 code 欄型別列成 DTYPE_CASES 釘住：可空 int/float、物件字串、category、
bool、全 NaN、±inf、超出 int64 範圍——每一種不是解得出來就是退回 1.0，都不 raise。

+++ src/recsys_tfb/io/extract.py
@@ -154,7 +154,15 @@ def decode_weight_keys(
     decoded = frame.copy()
     undecodable = np.zeros(len(frame), dtype=bool)
     for col, categories in decode_map.items():
-        raw = pd.to_numeric(frame[col], errors="coerce").to_numpy(dtype=np.float64)
+        # ``.astype("float64")`` before ``to_numpy``: a nullable dtype holding
+        # ``pd.NA`` (Int64 is reachable here — see ``dropna=False`` in
+        # :func:`_weight_keys_and_codes`) refuses to become a float64 numpy
+        # array directly, and the coerce alone does not widen it.
+        raw = (
+            pd.to_numeric(frame[col], errors="coerce")
+            .astype("float64")
+            .to_numpy(dtype=np.float64)
+        )
         # Clip before the int cast: casting inf or 1e30 to int64 is undefined,
         # and a clipped value fails the `safe == idx` identity below anyway.
         safe = np.clip(np.where(np.isfinite(raw), raw, -1.0), -1.0, float(len(categories)))
